Tue Jul 17 21:16:10 UTC 2018
Happy 25th anniversary to the Slackware 1.00 release! When the original
announcement went out on Usenet, I believe it had a UTC timestamp which has
led to some confusion over whether the anniversary falls on the 16th (which
was the date when I made the post) or on the 17th (which is when most people
first saw it)... but really, what's the difference? We can celebrate on both
days as far as I'm concerned. Thanks for sticking with the project all these
years. Glad I was able to help. :)
Here's a link to the 1.00 announcement:
  http://www.slackware.com/announce/1.0.php
And here's a nice article that was posted on opensource.com:
  https://opensource.com/article/18/7/stackware-turns-25
a/kernel-firmware-20180717_8d69bab-noarch-1.txz:  Upgraded.
l/pulseaudio-12.2-x86_64-1.txz:  Upgraded.
n/gnutls-3.6.3-x86_64-1.txz:  Upgraded.
n/mutt-1.10.1-x86_64-1.txz:  Upgraded.
  This update fixes bugs and security issues. Upstream strongly recommends
  that all IMAP and POP users upgrade as soon as possible.
  (* Security fix *)